move child logic to seperate file
